move io into serialize
